Send all CSV data to client
The client will need different kinds of data (game/shopping/travel) from currater.
It's better the server send all datat to the client.
So the client can decide what to do with that data instead of changing the server api.

But the server will still need to check the data validity. See #84